repositories change id type Long to Integer

+++ Backend/SkiStation/src/main/java/org/skistation/repositories/LiftRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LiftRepository extends JpaRepository<Lift, Long>
+public interface LiftRepository extends JpaRepository<Lift, Integer>
 {
 }
